pyproject.toml
[project]
name = "torch-projects"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Huvinesh Rajendran", email = "huvinesh@teleme.co" }]
dependencies = [
    "pytdc>=1.0.6",
    "torch>=2.4.0",
    "polars>=1.5.0",
    "numpy<2.0.0",
    "torchvision>=0.19.1",
    "rdkit-pypi>=2022.9.5",
    "networkx>=3.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
venvPath = "."
venv = ".venv"


[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/torch_projects"]
